Rename src directory to be "packages"
This is in preparation to make this a Lerna repo
