+++ packages/resourceful-redux/action-reducers.js
