+++ packages/resourceful-redux/action-types.js
