+++ packages/resourceful-redux/index.js
