+++ packages/resourceful-redux/reducers/create.js
