+++ packages/resourceful-redux/reducers/delete.js
