+++ packages/resourceful-redux/reducers/index.js
