+++ packages/resourceful-redux/reducers/read.js
