+++ packages/resourceful-redux/reducers/update.js
