+++ packages/resourceful-redux/resource-reducer.js
