+++ packages/resourceful-redux/utils/compose-reducers.js
