+++ packages/resourceful-redux/utils/generate-default-initial-state.js
