+++ packages/resourceful-redux/utils/get-resources.js
