+++ packages/resourceful-redux/utils/initial-resource-meta-state.js
