+++ packages/resourceful-redux/utils/reducer-generator.js
