+++ packages/resourceful-redux/utils/set-resource-meta.js
