+++ packages/resourceful-redux/utils/upsert-resources.js
