+++ test/unit/reducers/delete.js
@@ -1,4 +1,4 @@
-import {resourceReducer, requestStatuses} from '../../../src/resourceful-redux';
+import {resourceReducer, requestStatuses} from '../../../packages/resourceful-redux';
 
 describe('reducers: delete', function() {
   describe('DELETE_RESOURCES_SUCCEEDED', () => {
