+++ test/unit/reducers/read.js
@@ -1,4 +1,4 @@
-import {resourceReducer, requestStatuses} from '../../../src/resourceful-redux';
+import {resourceReducer, requestStatuses} from '../../../packages/resourceful-redux';
 
 describe('reducers: read:', function() {
   describe('READ_RESOURCES_SUCCEEDED:', () => {
