+++ test/unit/utils/update-meta-helper.js
@@ -1,5 +1,5 @@
-import reducerGenerator from '../../../src/resourceful-redux/utils/reducer-generator';
-import {requestStatuses} from '../../../src/resourceful-redux';
+import reducerGenerator from '../../../packages/resourceful-redux/utils/reducer-generator';
+import {requestStatuses} from '../../../packages/resourceful-redux';
 
 describe('reducerGenerator:', function() {
   it('passing no IDs and no label', () => {
